Update patch set 1
Patch Set 1:

New version of the original change that uses a stripped down copy of the zipalign code: https://android-review.googlesource.com/179560

It now rewrites the file headers in the original zip file instead of requiring a copy (a win on the giant packaging zip files). AFAICT libziparchive doesn't support that mode of operation yet.

Patch-set: 1