Update patch set 4
Patch Set 4:

(1 comment)

For non-boilerplate code, it's ~75 lines (not including comments/whitespace) in ZipFile.cpp that are mostly shared -- searching and reading the central directory 'headers'.

But it's not easily shared, especially since ziptime has no external dependencies, while zipalign has a bunch (and will be getting more as it is moved to using libziparchive). libziparchive's current API would mostly work for zipalign currently, but not ziptime -- ziptime rewrites the zip files in place, so it's not all passing through memory.

I expect this to go away at some point with soong. We've already replaced jar with soong_jar, I expect a similar tool for zip that never writes timestamps in the first place. But we don't have go build tool support in make right now, so that can't happen yet.

Patch-set: 4