Update patch set 1
Patch Set 1:

This should be fixed in zipalign as well.

This isn't a full zip library -- and it intentionally doesn't have any dependencies outside this repo. libziparchive doesn't support what's needed here anyway -- in place header rewriting.

Once we start depending on soong, I intend to remove this and use a go-based tool instead of using the system zip binary - that doesn't write timestamps in the first place.

Patch-set: 1