Update patch set 1
Patch Set 1: Code-Review-1

We do check in and distribute host tools built with this, so they should be reproducible.

It's definitely possible to configure gdb so that you don't need to be in $ANDROID_BUILD_TOP -- are you using plain gdb, or do we have a wrapper script for host tools?

Patch-set: 1
Reviewer: Dan Willemsen <1074628@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1