Update patch set 1
Patch Set 1:

I think David is just using plain gdb (although he worked around this with some configuration logic). Do we really want locally built host tools to be harder to debug though? I realize it is hard to know whether a buildbot or a person is doing the build, but maybe we should be differentiating here.

Patch-set: 1