Update patch set 1
Patch Set 1:

It's strange that there are different issues every time this is submitted... :\

I suspect that there are dependencies on the stack being zeroed and that the canaries can interfere. Otherwise, I would expect to actually see proper SSP errors if there's actually an out-of-bounds write hitting it.

Patch-set: 1