Update patch set 1
Patch Set 1:

> The camera code can opt-out of these protections, no?

yes, though i worry that if we do that for them, they'll never come back and fix their bugs. but it might be worth doing that anyway just so we can see what else is broken. if x86/x86-64 stick and there's still no progress, we could try that.

Patch-set: 1