Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

The event-log-tags.h file is dependent on what will be installed. So a package using a FOO_BAR_LOG_TAG define needs to depend on the modules that declare that in their logtags file.

But that doesn't even help with checkbuild, since you're building components that won't be installed.

So this probably should be a generated header per-module, and follow the normal exported c include scheme.

Patch-set: 1
Label: Code-Review=-1