Update patch set 1
Patch Set 1:

I don't disagree about the build time verification, but we don't want things like liblog to depend on a file that may change based on device or vendor specific things.

Is this just logd/liblog, or are there other instances where we need log tag values in C/C++? Should we just be reading it from the on-device database?

(The java generation is also somewhat broken -- it uses the global list, but falls back to a preset 999999 if it's not going to be installed. But at least it's per-module, so we shouldn't actually see a problem very often)

Patch-set: 1