Update patch set 1
Patch Set 1:

> Yeah, this won't help if anyone is using $(shell java*), since that
 > environment isn't modified, but it'll help. I suppose we could move
 > this up so that it's executed in the make instance instead of kati
 > -- that's a much smaller environment, and would expose these
 > modifications to kati itself, including $(shell). Thoughts Colin?

I'd rather not get too much logic in the makefiles outside kati, since eventually we want people to be able to build directly using soong to enable new features.  Kati stores a list of all the $(shell) commands it ran in the timestamp file, can you see if there are any that ran java in there?

Patch-set: 1