Update patch set 2
Patch Set 2: Code-Review+1

I'll bet that there might be some people out there explicitly building .jar files for host for things like vogar. We can't know, and I assume if they are they'll be able to switch to target equivalents. Might be worth holding on anything that builds on top of this change for a few days if you were planning it in case somebody needs a rollback.

The people to watch will be the ART team (who do a lot of work on host).

Patch-set: 2
Reviewer: Neil Fuller <1038039@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1