Update patch set 1
Patch Set 1: Code-Review-1

The mips64r6 build of Android is now broken on r11 prebuilts/ndk, whether this patch is added or not.  The patch makes it worse.  The only correct fix is to restore the presence of all lib/mips32r6 libraries in the r11 prebuilts, as was present in the aosp copy of prebuilts/ndk before the r11 switch over.

Patch-set: 1
Reviewer: Duane Sand <1080051@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1