Update patch set 2
Patch Set 2:

That's what this patch was looking to fix. The problem now is that we have NDK libraries being linked in to platform libraries (some third party libraries like libz and libgif are compiled only with the NDK, and thus for r1, but then linked into r6 platform code).

We never should have been in this situation in the first place. We shouldn't mix and match NDK and platform. There should be a libz and a libz_ndk like we do with gtest. Unfortunately since the problem hasn't had any symptoms until now, the issue is likely pretty widespread.

Patch-set: 2