Update patch set 2
Patch Set 2:

yeah, the internal structure of rangeset is a priority queue and it expects the input data to be sorted. Right now, I only store the blocks in file_map as it is while keep the block order everywhere else sorted. But we definitely need the original order in the file_map to compute imgdiff. Maybe we can have a unsorted copy in rangeset. And when doing operation (subtraction/union), we sort the data first.

Patch-set: 2