Update patch set 2
Patch Set 2:

> yeah, the internal structure of rangeset is a priority queue and it
 > expects the input data to be sorted. Right now, I only store the
 > blocks in file_map as it is while keep the block order everywhere
 > else sorted. But we definitely need the original order in the
 > file_map to compute imgdiff. Maybe we can have a unsorted copy in
 > rangeset. And when doing operation (subtraction/union), we sort the
 > data first.

(i don't see why it's a priority queue... looks like a regular sorted list to me)

things like rangelib.RangeSet("15-20 30 1-5", False).subtract(rangelib.RangeSet("30")) are valid cases that we must handle properly.

why don't we go the other way around? i.e. to keep the data as sorted internally, but also keeping a copy of the input string if it's out-of-order. when outputting the string representation, we then follow the order in the input string.

Patch-set: 2