Update patch set 3
Patch Set 3:

In the context of this tool, just generate the groups and passwd files to match the oem header content. If not comfortable with C++, I will work on the parser in grp_pwd.cpp when I have time.

Patch-set: 3
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0