Update patch set 2
Patch Set 2:

In fact, soong ignores the per-target sanitize property in a global sanitizer is set. This is different from the make build system, which takes the set union of local and global sanitizers.

Is it intentional?

At this time, the only visible difference is that SANITIZE_TARGET=address soong build will miss the integer overflow sanitization for the modules where it is enabled.

Patch-set: 2