Update patch set 1
Patch Set 1:

Current AOSP build default does not turn on clang-tidy, static analysis, or Error Prone, right? If any of them is on by default, many people will ask to suppress those warnings with thousands of cases at compile time, not just in warn.py.

So there is no need to have a hard coded warning limit here.
How about just a flag for you, me, and anyone to use in special builds?
I would prefer the default to be no limit.
The limit of 1000 is still too low for clang-tidy.
I will need a flag to unset that limit for my nightly runs that check for tens of thousands of warnings.

If you are sure that only "MissingOverride" is causing the trouble,
another choice is to have a limit of that specific kind.

Patch-set: 1