Update patch set 1
Patch Set 1:

> Current AOSP build default does not turn on clang-tidy, static
 > analysis, or Error Prone, right? If any of them is on by default,
 > many people will ask to suppress those warnings with thousands of
 > cases at compile time, not just in warn.py.
 
Irrelevant, generating a too big HTML file is still broken.

 > So there is no need to have a hard coded warning limit here.

There is, so that generated HTML files aren't too big.

 > How about just a flag for you, me, and anyone to use in special
 > builds?

Its a nuisance.

 > I would prefer the default to be no limit.
 > The limit of 1000 is still too low for clang-tidy.

This doesn't limit the total number of warnings processed, just those that are placed in the generated HTML files. So count numbers are accurate.

 > I will need a flag to unset that limit for my nightly runs that
 > check for tens of thousands of warnings.
 
Or look at the build.log. This is a maximum limit for a particular class of warning, not total number of warnings in the HTML file.

 > If you are sure that only "MissingOverride" is causing the trouble,
 > another choice is to have a limit of that specific kind.

Currently I can only disable warnings and remembering to reenable a useful warning requires attention that it's not clear will be given. I specifically want to work around the large file size problem that takes nothing a way from anything you are saying.

Patch-set: 1
Reviewer: Ian Rogers <1010118@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0