Update patch set 1
Patch Set 1:

This would disable ubsan on ndk code, as well. I don't know if any ndk code is built with ubsan, but we might want to do this one day.

Maybe it would be better to disable it only for the sanitizers that have runtime libraries, i.e. address and thread?

Patch-set: 1