Update patch set 2
Patch Set 2:

I've tweaked -fsanitize-trap flags to work around the fact that our libclang_rt.asan does not include ubsan handlers, so, even with asan, we want to build ubsan in the trapping mode. This is temporary until the asan runtime library is fixed.

As for the LOCAL_SANITIZE=address targets, we can worry about them later. As you said, it should be a lot easier with soong.

Patch-set: 2