Update patch set 1
Patch Set 1:

This will simplify updating clang, especially with the version currently hardcoded in make and soong, but does have a disadvantage - changing the symlink may not cause anything to rebuild as the command line would not change, and the clang binary at the new location may have the same or older timestamp than the one at the old location.

We could use $(shell readlink) to evaluate the symlink at make parsing time here, and kati would rerun any time the result of that changed, but there is currently no way to depend on a symlink in soong to know when to rebuild the ninja file with a new version.

Patch-set: 1
Reviewer: Colin Cross <1002751@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=0