Update patch set 1
Patch Set 1:

I'm not sure that I like the idea of disabling sanitization for NDK compiles in the platform build. There are legitimate cases for debugging apps and/or the platform where we should have a way to compile with ASan, even if the library is using the NDK. As it stands, I would now have to port that code to use the regular NDK and then push it onto the device (or revert this change). Are you just trying to prevent people enabling sanitizers by default? Why not check that it was done at build/configuration time, rather than by a stray Makefile?

Patch-set: 1