Update patch set 1
Patch Set 1:

> I think that WITHOUT_COMPILERRT will affect both AUX_ and TARGET_
 > components, while this option is only suppressing it for AUX_ but
 > not TARGET_

So then this is indeed building the rest of a target. Why would compiler-rt not be there for the rest of the device?

 > I think that WITHOUT_COMPILERRT will affect both AUX_ and TARGET_
 > components, while this option is only suppressing it for AUX_ but
 > not TARGET_

If that is the case, then won't compiler-rt need to be there for the rest of the device? Is this a trimmed down target? I guess I am failing to see how this actually breaks your build, considering any functions here will likely go completely unused. I will leave it to Dan to ultimately decide here, because this tangled mess of build/ is not something that I would want, but it is his decision. As I said, you can leave the static library and it should be harmless (and in the future, it will just disappear).

Patch-set: 1