Update patch set 1
Patch Set 1:

On ARM:
      0x71acdb50: 9803          ldr     r0, [sp, #12]
      0x71acdb52: f24521e2      movw    r1, #21218
      0x71acdb56: f6cf61a0      movt    r1, #65184
      0x71acdb5a: 4479          add     r1, pc
      0x71acdb5c: f9b21000      ldrsh.w r1, [r2, #0]
      0x71acdb60: b100          cbz     r0, +0 (0x71acdb64)
      0x71acdb62: bac9          revsh   r1, r1
      0x71acdb64: 4608          mov     r0, r1
      0x71acdb66: 4770          bx      lr
The "MOVW+MOVT+ADD r1,pc" is a dead HArmDexCacheArraysBase and it is harmless (except for a performance hit).

However, on MIPS, the dead HMipsComputeBaseMethodAddress clobbers the RA.

We need to fix both ARM and MIPS to not add the special input to intrinsified HInvokeStaticOrDirect.

Patch-set: 1
Reviewer: Vladimir Marko <1018108@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=0