Update patch set 1
Patch Set 1:

> What has changed?

The comments on the bug are unanimous that the Google JDK team see no reason for their JDK to be black listed. This was last updated in May. I'm uploading this now as I needed the CL in a different tree. Carrying this change is inconvenient for me. If you could update the bug explaining why you think the Google JDK is inappropriate for the Android build then I think it would be appropriate to abandon this change. Thanks!

Patch-set: 1
Reviewer: Ian Rogers <1010118@85c56323-6fa9-3386-8a01-6480fb634889>