Update patch set 1
Patch Set 1:

Can you point me to your Android.mk? There are often other problems with dealing with multiple files in the same rule. We've got $(call copy-many-files) to handle some of the problems.

If we were to change these, we should change all of them, not just this version.

Patch-set: 1