Update patch set 2
Patch Set 2:

Colin, i'm still learning soong.
Since soong is in a different project, the changes will be in another CL.
Will it be okay to merge in this one first and get it tested in all branches and targets?

Patch-set: 2
Reviewer: Chih-hung Hsieh <1052399@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0