Update patch set 1
Patch Set 1: Code-Review-2

The point of doing LOCAL_JACK_VERSION differently is to hide it and avoid its declaration in any Android.mk. LOCAL_JACK_VERSION is just a way to experiment future release of Jack or to workaround some problem we want it to be controlled with the releases of Jack in prebuilts/sdk/tools and not to be set in any Android.mk.

Patch-set: 1
Label: Code-Review=-2