Update patch set 1
Patch Set 1:

In order to use LOCAL_PROPRIETARY_MODULE properly, you'd also need to use LOCAL_MODULE_OWNER.

But we really don't want to keep growing this list -- it's not maintainable. The suggestion is to use BUILD_PREBUILT instead of BUILD_MULTI_PREBUILT, and if you actually need to loop, do it in your Android.mk file.

Patch-set: 1