Update patch set 1
Patch Set 1:

Is there any way to have the build system tell TreeHugger what target failed when it failed?

You could package the build into three stages (e.g. build all first, then tests, then checkbuild), but it's liable to be even slower that way. What we'd want is to build all three together, but get a signal back from make which target failed and what it belonged to.

Keep in mind, you're making developer builds faster with this, but you're also testing less -- you're saying that 'tests are less important than other code and don't need to compile in order for your change to work locally'. Is that the mindset we want to promote? 

As we add more tests the probability of tests failing to compile is going to grow, and I wonder if this change is going to cost more time 
than it saves as developers get frustrated with uploading their change and finding out 2 hours later that it fails to compile. Something to think about.

I'm not arguing against this change -- just want to make sure we've discussed the implications widely and developed a plan for addressing them. Perhaps starting a doc for this would help and we can set up a plan for addressing all the items different teams will need to work on to make this possible?

Patch-set: 1
Reviewer: David James <1076194@85c56323-6fa9-3386-8a01-6480fb634889>